% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_parameters.R
\name{initialize_parameters}
\alias{initialize_parameters}
\title{initialize a list of parameters}
\usage{
initialize_parameters(title, survey.file, choices.file, choices.label,
  survey.label, right.to.left, special.characters)
}
\arguments{
\item{title}{Name of your questionnaire as character.}

\item{survey.file}{Path to your questionnaire csv file as character.}

\item{choices.file}{Path to your choices csv file as character.}

\item{choices.label}{Choices label column to be used as character.}

\item{survey.label}{questionnaire label column to be used as character.}

\item{right.to.left}{Type of the questionnaire writing system (TRUE/FALSE).}

\item{special.characters}{If the text contains non latin characters, specify the language used.}
}
\value{
list of 4 objects.
}
\description{
initialize a list of parameters
}
\examples{
initialize_parameters("Area Based Assessment 2019", "questionnaire_file.csv", "choices_file.csv", "label::Arabic",
"label::Arabic", TRUE, "arabic")

initialize_parameters("JMMI January round", "questionnaire_file.csv", "choices_file.csv", "label",
"label", FALSE, "")
}
