% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Create a list of parameters (inputs)}
\usage{
load_parameters(title, survey.file, choices.file, choices.label,
  survey.label, right.to.left, special.characters)
}
\arguments{
\item{title}{Name of your questionnaire as character.}

\item{survey.file}{Path to your questionnaire csv file as character.}

\item{choices.file}{Path to your choices csv file as character.}

\item{choices.label}{Choices label column to be used as character.}

\item{survey.label}{questionnaire label column to be used as character.}

\item{right.to.left}{Type of the questionnaire writing system (TRUE/FALSE).}

\item{special.characters}{If the text contains non latin characters, specify the language used.}
}
\value{
list of 4 objects.
}
\description{
Create a list of parameters (inputs)
}
\examples{
load_parameters("Area Based Assessment 2019", "./input/questionnaire_file.csv", "./input/choices_file.csv", "label::Arabic",
"label::Arabic", TRUE, "arabic")

load_parameters("JMMI January round", "./input/questionnaire_file.csv", "./input/choices_file.csv", "label",
"label", FALSE, "")
}
