% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_questionnaire_without_data.R
\name{load_questionnaire_without_data}
\alias{load_questionnaire_without_data}
\title{Using the function Load questionnaire from koboquest with fake data}
\usage{
load_questionnaire_without_data(questions.file, choices.file,
  choices.label.column.to.use = NULL)
}
\arguments{
\item{questions.file}{Path to your questionnaire csv file as character.}

\item{choices.file}{Path to your choices csv file as character.}

\item{choices.label.column.to.use}{Choices label column to be used as character.}
}
\value{
Loaded questionnaire
}
\description{
Using the function Load questionnaire from koboquest with fake data
}
\examples{
load_questionnaire_without_data("questionnaire_file.csv", "choices_file.csv", "label::Arabic",)
}
