% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/questionnaire_to_html.R
\name{questionnaire_to_html}
\alias{questionnaire_to_html}
\title{Convert an XLSform questionnaire to HTML}
\usage{
questionnaire_to_html(survey.file, choices.file, choices.label,
  survey.label, output.dir = "./",
  output.filename = "questionnaire.html", title = "",
  right.to.left = FALSE, special.characters = "")
}
\arguments{
\item{survey.file}{Path to your questionnaire csv file as character.}

\item{choices.file}{Path to your choices csv file as character.}

\item{choices.label}{Choices label column to be used as character.}

\item{survey.label}{questionnaire label column to be used as character.}

\item{output.dir}{the directory in which to save the output file (absolute path or relative to current working directory)}

\item{output.filename}{the name of the file. must end in '.html'}

\item{title}{Name of your questionnaire as character.}

\item{right.to.left}{Type of the questionnaire writing system (TRUE/FALSE).}

\item{special.characters}{If the text contains non latin characters, specify the language used.}
}
\description{
Convert an XLSform questionnaire to HTML
}
\examples{
questionnaire_to_html("./input/questionnaire_file.csv", "./input/choices_file.csv", "label::Arabic",
"label::Arabic", "./output", "survey.html", "Area Based Assessment 2019", TRUE, "arabic")

questionnaire_to_html("./input/questionnaire_file.csv", "./input/choices_file.csv", "label",
"label", "./", "questionnaire.html", "Area Based Assessment 2019", FALSE, "")
}
