% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_questionnaire_rmd.R
\name{render_questionnaire_rmd}
\alias{render_questionnaire_rmd}
\title{Map results to an output template}
\usage{
render_questionnaire_rmd(x, dir, filename)
}
\arguments{
\item{x}{list of parameters}

\item{dir}{the directory in which to save the output file (absolute path or relative to current working directory)}

\item{filename}{the name of the file. must end in '.html'}
}
\description{
Map results to an output template
}
