% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_choices_label_listname.R
\name{sanitise_choices_label_listname}
\alias{sanitise_choices_label_listname}
\title{Rename the header used in the choices file to label}
\usage{
sanitise_choices_label_listname(original.label, choices.names)
}
\arguments{
\item{original.label}{Name of the survey header to use.}

\item{choices.names}{names of all columns.}
}
\description{
Rename the header used in the choices file to label
}
