% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sanitise_survey_label.R
\name{sanitise_survey_label}
\alias{sanitise_survey_label}
\title{Rename the header used in the questonnaire file to label}
\usage{
sanitise_survey_label(original.label, survey.names)
}
\arguments{
\item{original.label}{Name of the survey header to use.}

\item{survey.names}{names of all columns.}
}
\description{
Rename the header used in the questonnaire file to label
}
