% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xlsform_to_html.R
\name{xlsform_to_html}
\alias{xlsform_to_html}
\title{Map results to an output template}
\usage{
xlsform_to_html(x, dir, filename)
}
\arguments{
\item{x}{list of parameters (created with initialize_parameters)}

\item{dir}{the directory in which to save the output file (absolute path or relative to current working directory)}

\item{filename}{the name of the file. must end in '.html'}
}
\description{
Map results to an output template
}
